backend/requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-cors==4.0.0
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt==4.1.1
google-genai==0.3.0
